refactor: move AuthGuard to src/ and add query provider
